Fix leveldb resource leak (#273)
`createStorage` can return before the child goroutine that opens the db
resource, leading to a resource leak that prevents subsequent attempts
to open the same db from succeeding.

Considering that the intent of the concurrency in the scope of
`createStorage` is to print notifications, it is more robust to do the
notifications in a child goroutine and do the resource handling in the
parent where it's simpler to block and cleanup if necessary.