fix: Kafka client resource leak
Sarama Kafka clients leak several resources (goroutines, metrics,
connections) if not closed properly. A client is created and managed
within the topic manager and therefore a topic manager will leak
resources if not properly closed.

Topic Managers are created in two places:

1. When calling `prepareTopics` while instantiating a new processor.
2. Within `processor.Run`.

The first of these is correctly closed, but the latter was not. Closing
the second instance should ensure underlying Sarama resources are
appropriate cleaned up when a processor stops.